requirements.txt
streamlit
pandas
matplotlib
seaborn
openpyxl
sentence-transformers
chromadb
requests
python-dotenv